Refactor BlockEncoding: Base Class + Facade Pattern
